Add support for retrieving original vectors passed to psa_call()
New config option CONFIG_TFM_VECTOR_ACCESS that defaults to 0 (off).

In some cases, for example dcache operations, the content of an address
block is unimportant, but the existing accessibility checks are. In
these cases, it is necessary to retrieve the base address and size as
passed in to psa_call().

Co-Authored-by: Roman Mazurak <roman.mazurak@infineon.com>
Co-Authored-by: Bohdan Hunko <Bohdan.Hunko@infineon.com>
Signed-off-by: Chris Brand <chris.brand@cypress.com>
Change-Id: I3820b1d9e917f56d84e9edef0a70f393322b7c46
(cherry picked from commit ff2806fd919fd0addcedd9b852a8a2dd26e4afc7)